[GUI] Split gui widgets into own header files
