+++ src/client/gui/widget/image_widget.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "widget.h"
+
+namespace gui {
+/**
+ * @brief Raw image overlay widget
+ */
+class ImageWidget final : public Widget {
+  public:
+    ImageWidget(RectangleComponent* rectangleComponent);
+    void setPosition(const GuiDimension& position) final override;
+    void setSize(const GuiDimension& size) final override;
+
+    void setImage(int image);
+
+  private:
+    RectangleComponent* mp_rectangle;
+};
+} // namespace gui
